Moved constants to Constants + Extracted BepEventHandler from BepServer
